New option for Settings page for 'recaptchacompat'
Allows adding a URL parameter to stop the replacement of reCaptcha outside of the plugin's included modules.
Use case: I have sites that include a Constant Contact signup widget, and they include reCaptcha v3 invisible by default with no way to manage the sitekey or secret. Adding this option and enabling it on these sites allows both types of captcha to work properly, even on the same page.